# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Legível",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/capture/check": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Check Capture",
        "operationId": "check_capture_api_capture_check_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_check_capture_api_capture_check_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Check Capture Api Capture Check Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "Jobs",
        "operationId": "jobs_api_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobOut"
                  },
                  "type": "array",
                  "title": "Response Jobs Api Jobs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Enqueue",
        "operationId": "enqueue_api_jobs_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_enqueue_api_jobs_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_api_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/retry": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Retry Job",
        "operationId": "retry_job_api_jobs__job_id__retry_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organization": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "Organized",
        "operationId": "organized_api_organization_get",
        "parameters": [
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Folder"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Organized Api Organization Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organization/{entity}/{entity_id}": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "Organization",
        "operationId": "organization_api_organization__entity___entity_id__get",
        "parameters": [
          {
            "name": "entity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Organization Api Organization  Entity   Entity Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "workflow"
        ],
        "summary": "Save Organization",
        "operationId": "save_organization_api_organization__entity___entity_id__put",
        "parameters": [
          {
            "name": "entity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Save Organization Api Organization  Entity   Entity Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/saved-searches": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "Saved Searches",
        "operationId": "saved_searches_api_saved_searches_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Saved Searches Api Saved Searches Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Save Search",
        "operationId": "save_search_api_saved_searches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Save Search Api Saved Searches Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/saved-searches/{search_id}": {
      "delete": {
        "tags": [
          "workflow"
        ],
        "summary": "Delete Search",
        "operationId": "delete_search_api_saved_searches__search_id__delete",
        "parameters": [
          {
            "name": "search_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Search Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/redaction/{entity}/{entity_id}": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Export Redacted",
        "operationId": "export_redacted_api_redaction__entity___entity_id__post",
        "parameters": [
          {
            "name": "entity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedactionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_check_capture_api_capture_check_post": {
        "properties": {
          "photo": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Photo"
          }
        },
        "type": "object",
        "required": [
          "photo"
        ],
        "title": "Body_check_capture_api_capture_check_post"
      },
...
      "Body_enqueue_api_jobs_post": {
        "properties": {
          "module": {
            "type": "string",
            "title": "Module",
            "default": "documents"
          },
          "front": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Front"
          },
          "back": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Back"
          },
          "pages": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "contentMediaType": "application/octet-stream"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": ""
          },
          "duplicate": {
            "type": "string",
            "enum": [
              "ask",
              "keep",
              "link",
              "replace"
            ],
            "title": "Duplicate",
            "default": "ask"
          }
        },
        "type": "object",
        "title": "Body_enqueue_api_jobs_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "module": {
            "type": "string",
            "title": "Module"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "completed_pages": {
            "type": "integer",
            "title": "Completed Pages"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "result_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "module",
          "label",
          "status",
          "completed_pages",
          "total_pages",
          "result_id",
          "error",
          "created_at"
        ],
        "title": "JobOut"
      },
...
      "OrganizationIn": {
        "properties": {
          "folder": {
            "type": "string",
            "maxLength": 100,
            "title": "Folder",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "OrganizationIn"
      },
...
      "RedactionIn": {
        "properties": {
          "regions": {
            "items": {
              "$ref": "#/components/schemas/RedactionRegion"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Regions"
          },
          "preview_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Page"
          }
        },
        "type": "object",
        "required": [
          "regions"
        ],
        "title": "RedactionIn"
      },
      "RedactionRegion": {
        "properties": {
          "page_id": {
            "type": "integer",
            "title": "Page Id"
          },
          "rect": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Rect"
          }
        },
        "type": "object",
        "required": [
          "page_id",
          "rect"
        ],
        "title": "RedactionRegion"
      },
...
      "SavedSearchIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "path": {
            "type": "string",
            "maxLength": 2000,
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "name",
          "path"
        ],
        "title": "SavedSearchIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}